+++ src/main/java/org/bot/nullbot/command/ai/embedding/BanChatCommand.java
@@ -30,12 +30,12 @@ public class BanChatCommand implements Command
         }else if(event.getEvent() instanceof PokeNoticeEvent pokeNoticeEvent){
             groupId = pokeNoticeEvent.getGroupId();
         }else{
-            log.info("\t\t\t\t├─[Meme] 未设计 非群消息/戳一戳事件响应方式");
+            log.info("\t\t\t\t├─[BanChat] 未设计 非群消息/戳一戳事件响应方式");
             return;
         }
 
         if(event.getCommandParameters().size() < 2) {
-            bot.sendGroupMsg(groupId, "[封禁对话] ❌参数不足", false);
+            bot.sendGroupMsg(groupId, "[封禁AI对话] ❌参数不足", false);
             log.info("\t\t\t\t├─[BanChat] 参数不足");
             return;
         }
@@ -44,15 +44,15 @@ public class BanChatCommand implements Command
             long userId = Long.parseLong(event.getCommandParameters().get(0));
             int time = Integer.parseInt(event.getCommandParameters().get(1));
             if(!userService.existUser(userId)) {
-                bot.sendGroupMsg(groupId, "[封禁对话] ❌用户不存在", false);
+                bot.sendGroupMsg(groupId, "[封禁AI对话] ❌用户不存在", false);
                 log.info("\t\t\t\t├─[BanChat] 用户不存在");
             }
             chatStorage.banUser(userId, time);
             String userName = bot.getStrangerInfo(userId, true).getData().getNickname();
-            bot.sendGroupMsg(groupId, "[封禁对话] ✅已设置！\n" + userName + "(" + userId + ") -> " + time + " Min", false);
+            bot.sendGroupMsg(groupId, "[封禁AI对话] ✅已设置！\n" + userName + "(" + userId + ") -> " + time + " Min", false);
             log.info("\t\t\t\t├─[BanChat] 已封禁对话 - {} -> {} min", userId, time);
         } catch (NumberFormatException e) {
-            bot.sendGroupMsg(groupId, "[封禁对话] ❌参数格式错误", false);
+            bot.sendGroupMsg(groupId, "[封禁AI对话] ❌参数格式错误", false);
             log.info("\t\t\t\t├─[BanChat] 参数格式错误");
         }
     }